# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "CD-PIPELINE-SKELETON",
    "description": "Skeleton for a cd node project"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "Example",
      "description": "Example API endpoints"
    },
    {
      "name": "Questions",
      "description": "Question APIs"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/example": {
      "get": {
        "tags": [
          "Example"
        ],
        "summary": "Example get",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/questions": {
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "Returns all different questions",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "ERROR"
          }
        }
      }
    },
    "/questions/random": {
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "Returns a random question",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "ERROR"
          }
        }
      }
    },
    "/questions/answer": {
      "post": {
        "tags": [
          "Questions"
        ],
        "summary": "Returns true if the given answer was correct",
        "parameters": [
          {
            "in": "body",
            "name": "id",
            "description": "id of the question that was answered",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "body",
            "name": "answer",
            "description": "user's answer",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ isCorrect: true/false }"
          },
          "500": {
            "description": "ERROR"
          }
        }
      }
    }
  }
}